add review package with utilities for text processing, image verification, and database interactions
