+++ createArticles/review/text_utils.py
@@ -1,0 +1,23 @@
+"""
+Text utility functions for cleaning and processing text content.
+"""
+
+import re
+
+def clean_text(text: str) -> str:
+    """Remove all '\n' sequences, clean up spacing, and remove leading quotes."""
+    if not text:
+        return text
+    
+    # Replace literal '\n' sequences with spaces
+    cleaned = text.replace('\\n', ' ')
+    # Replace actual newlines with spaces
+    cleaned = cleaned.replace('\n', ' ')
+    # Replace multiple spaces with a single space
+    cleaned = re.sub(r'\s+', ' ', cleaned)
+    # Remove leading quotes and space after quote if present
+    cleaned = re.sub(r'^[""]\s*', '', cleaned)
+    # Final trim
+    cleaned = cleaned.strip()
+    
+    return cleaned
